Use a fixture helper
