+++ server/app/tests/gagarin/testAppsPublications.js
@@ -1,30 +1,16 @@
+var fixtures = require('../helpers/fixtures.js');
+
 describe('Apps publications', function() {
-  var server = meteor();
+  var server = meteor({
+    helpers: fixtures
+  });
   var client = browser(server);
 
   var sampleAppId;
   var sampleAppAppId;
 
   before(function() {
-    return server.execute(function(sampleAppId) {
-      var sampleApp = new App({
-        name: 'Sample App',
-        metrics: {
-          'some-user-1': {
-            'some-metric-1': 42,
-            'some-metric-2': 100
-          },
-          'some-user-2': {
-            'some-metric-3': 12,
-          }
-        }
-      });
-      var res = {};
-      res.id = sampleApp.save();
-      sampleApp.reload();
-      res.appId = sampleApp.get('appId');
-      return res;
-    }).then(function(res) {
+    return server.createSampleApp().then(function(res) {
       sampleAppId = res.id;
       sampleAppAppId = res.appId;
     });
