+++ server/app/tests/gagarin/testBrandFlakesJS.js
@@ -1,24 +1,17 @@
+var fixtures = require('../helpers/fixtures.js');
+
 describe('brand-flakes.js deliverable', function() {
-  var server = meteor();
+  var server = meteor({
+    helpers: fixtures
+  });
   var client;
 
   var sampleAppAppId;
 
   before(function() {
     // Create a sample app and snag its appId
-    return server.execute(function() {
-      var sampleApp = new App({
-        name: 'Sample App',
-        metrics: {
-          'some-user-id': {
-            'metric-to-get': 42
-          }
-        }
-      });
-      sampleApp.save();
-      return sampleApp.get('appId');
-    }).then(function(id) {
-      sampleAppAppId = id;
+    return server.createSampleApp().then(function(res) {
+      sampleAppAppId = res.appId;
     });
   });
 
@@ -27,26 +20,26 @@ describe('brand-flakes.js deliverable', function() {
     return server.execute(function() {
       return process.env.ROOT_URL;
     }).then(function(root_url) {
-      client = browser(root_url + '/?appId=' + sampleAppAppId + '&userId=some-user-id', {}, server);
+      client = browser(root_url + '/?appId=' + sampleAppAppId + '&userId=some-user-1', {}, server);
     });
   });
 
   it('can call record', function(){
     return client.execute(function() {
-      BrandFlakes.record('some-metric');
+      BrandFlakes.record('some-metric-3');
     }).then(function() {
       return server.wait(2000, 'for record to propogate to server', function() {
         var app = Apps.findOne();
-        return app.metrics['some-user-id'];
+        return app.metrics['some-user-1'];
       }).then(function(user) {
-        expect(user['some-metric']).to.equal(1);
+        expect(user['some-metric-3']).to.equal(1);
       });
     });
   });
 
   it('can call get', function(){
     return client.execute(function() {
-      return BrandFlakes.get('metric-to-get');
+      return BrandFlakes.get('some-metric-1');
     }).then(function(metric) {
       expect(metric).to.equal(42);
     });
